made the router cleaner
